@@ -3,9 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "prop-types": "^15.6.2",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
+    "prop-types": "^15.8.1",
+    "react": "^16.7.3",
+    "react-dom": "^16.7.3",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "2.1.1"
   },
   "scripts": {
